Fix lazy-instance template to preserve object alignment on MIPS.
The alignment attribute of an object was discarded when allocating storage
for the lazy-instance. Alignment is important for MIPS, where doubles
(and objects containing them) must be aligned on 8-byte boundary.

The alignment attributes are included via a macro that is only defined
under GCC, so that this change does not affect other build enviromments.

Change-Id: I99047cf610618de2e7e29e5de85835dc45d99231
Signed-off-by: Paul Lind <plind@mips.com>